brcmfmac: Return time since boot when reporting scan results.
The Android framework expects scan results to have the time since boot.
This value is compared with the timestamp of the scan request in the
function pollLatestScanData in
com.android.server.wifi.scanner.SupplicantWifiScannerImpl. The framework
only considers the scan results to be valid if the timestamp of the scan
is greater than that of the scan request.

BUG=31785785
TEST=rpi3 wifi reassociates on reboot

Change-Id: I8032a697f9cae42f605416abc55f96ff95b447fc